♻️ refactor: 리뷰 반영 - /infra/api/* 에는 요청/응답 타입만 관리
