+++ src/infra/api/meme/index.ts
@@ -1,11 +1,13 @@
 import type { AxiosInstance } from "axios";
 
-import type { Meme } from "./types";
+import type { GetMemeDetailByIdResponse } from "./types";
 
 export class MemeApi {
   constructor(private api: AxiosInstance) {}
 
   getMemeDetailById = (id: string) => {
-    return this.api.get<Meme>(`/memes/${id}`).then((response) => response.data);
+    return this.api
+      .get<GetMemeDetailByIdResponse>(`/memes/${id}`)
+      .then((response) => response.data);
   };
 }
